More realistic benchmarks
The *_map_* benchmarks try to emulate the same amount of insert/deletes
as in a LRU cache. While the deleted items are not necessarily the
oldest, the target hit/miss ratio is reached.

Also replaced calls to rand/v2 by an xorshift64* PRNG. Fast enough to
be usable in benchmarks without skewing the timings (~3 ns/op on an
AMD FX6300), and good enough for our purpose.