Re-add options and custom hash functions
capacity and custom hash functions are now available as Options.
Removed github.com/dolthub/maphash as the default hasher in favor of
the standard library's maphash.Hash. This is slower but does not rely on
internal APIs of the Go runtime (which are not guaranteed to be stable
from one Go version to the next). Also, the specizlized hashers provided
by the hash package are just as fast (for string and []byte) or much
faster for integers.